Collapse xtask OS-naming ladders into policy tables
mpv library filenames and package archive slug/extension each repeated
the same per-OS conditional across multiple functions. Consolidate the
naming into single-source-of-truth constants, leaving only the genuine
mechanism splits (archive writer) as branches.
